Fix wildcard character in regex strings.
The map directive contains regexes (strings starting with a tilde) that
contain 'asterisk' characters. In a regex, the asterisk character is a
metacharacter, and in this context, it is meant as a wildcard character.
To achieve the desired result, we should place a dot in front of the
asterisk character.